Moving util funcitons to task util
